requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
requests==2.32.0
httpx==0.27.0
sentence-transformers==3.0.0
faiss-cpu==1.8.0
lxml==5.3.0
numpy
